[#895] improve code based on sonar and review
