+++ frontend/styles/paper-corrected.css
@@ -16,6 +16,6 @@
   color: var(--blue-700);
 }
 
-.ex-officio-value {
+.csb-corrected-value {
   color: var(--red-700);
 }
