+++ src/csb/components/paper_corrected.html
@@ -1,8 +1,8 @@
 {% macro value(v) %}
-  {% if let Some(eo) = v.ex_officio %}
+  {% if let Some(csb_corrected) = v.csb_corrected %}
     {% if !v.imported.is_empty() %}<s class="imported-value">{{ v.imported }}</s>{% endif %}
     {% if v.differs() && !v.corrected.is_empty() %}<s class="imported-value">{{ v.corrected }}</s>{% endif %}
-    <strong class="ex-officio-value">{{ eo }}</strong>
+    <strong class="csb-corrected-value">{{ csb_corrected }}</strong>
   {% else if v.differs() %}
     {% if !v.imported.is_empty() %}<s class="imported-value">{{ v.imported }}</s>{% endif %}
     {% if !v.corrected.is_empty() %}<strong class="paper-corrected-value">{{ v.corrected }}</strong>{% endif %}
