+++ src/csb/examination/pages/candidate.rs
@@ -40,7 +40,7 @@ pub async fn overview(
     let corrected_store = store.paper_corrected();
     let imported = store.get_imported_person(person_id);
     let corrected = corrected_store.get_person(person_id).ok();
-    let ex_officio = store.get_corrected_person(person_id);
+    let csb_corrected = store.get_csb_corrected_person(person_id);
     let candidate = imported
         .clone()
         .or_else(|| corrected.clone())
@@ -48,7 +48,7 @@ pub async fn overview(
     let details = PaperCorrectedPersonDetails::new(
         imported.as_ref(),
         corrected.as_ref(),
-        ex_officio.as_ref(),
+        csb_corrected.as_ref(),
         context.session.locale,
     );
     let position = PaperCorrected::new(
