+++ src/csb/examination/pages/correction.html
@@ -39,7 +39,7 @@
           </p>
           {% when CorrectionFieldType::DateOfBirth %}
           <p class="form-field form-field-md">
-            <label for="date_of_birth">{{ "csb.correction.form.value_label"|trans }}</label>
+            <label for="day_of_birth">{{ "csb.correction.form.value_label"|trans }}</label>
             <span class="date-input-row">
               <span class="date-separate hidden">
                 <input type="text"
