+++ src/csb/examination/structs/paper_corrected/political_group_info.rs
@@ -19,7 +19,7 @@ impl PaperCorrectedPoliticalGroupInfo {
     pub fn new(
         store: &CsbStore,
         corrected: &PgStore,
-        ex_officio_display_name: Option<String>,
+        csb_corrected_display_name: Option<String>,
         locale: Locale,
     ) -> Self {
         let imported_group = store.get_imported_political_group();
@@ -30,7 +30,7 @@ impl PaperCorrectedPoliticalGroupInfo {
                 imported_group.csb_display_name(store.first_imported_candidate_name().as_ref()),
                 corrected_group.csb_display_name(corrected.first_candidate_name().as_ref()),
             )
-            .with_ex_officio(ex_officio_display_name),
+            .with_csb_correction(csb_corrected_display_name),
             list_type: PaperCorrected::new(
                 list_type_label(&imported_group, locale),
                 list_type_label(&corrected_group, locale),
